requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
cycler==0.11.0
facenet==1.0.5
flatbuffers==23.5.26
fonttools==4.38.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.56.2
h5py==2.10.0
idna==3.4
imageio==2.31.1
importlib-metadata==6.7.0
imutils==0.5.3
joblib==1.3.1
keras==2.11.0
kiwisolver==1.4.4
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
numpy==1.21.6
oauthlib==3.2.2
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
opt-einsum==3.3.0
packaging==23.1
pandas==1.3.5
Pillow==9.0.1
protobuf==3.19.6
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==2.3.0
threadpoolctl==3.1.0
typing-extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
